Rename WysiwygTiptap component file name to match Camelcase current naming convention
