+++ frontend/js/components/WysiwygTiptap.vue
