+++ frontend/js/plugins/A17Config.js
@@ -11,7 +11,7 @@ import a17ColorField from '@/components/ColorField.vue'
 import a17Textfield from '@/components/Textfield.vue'
 import a17HiddenField from '@/components/HiddenField.vue'
 import a17Wysiwyg from '@/components/Wysiwyg.vue'
-import a17WysiwygTipTap from '@/components/Wysiwyg-tiptap.vue'
+import a17WysiwygTipTap from '@/components/WysiwygTiptap.vue'
 import a17MediaField from '@/components/MediaField.vue'
 import a17MediaFieldTranslated from '@/components/MediaFieldTranslated.vue'
 import a17Radio from '@/components/Radio.vue'
